add support for config plugins to dynamically adjust the running config
This adds a new ConfigPlugin class, which can be used to return a dynamic
config block.  To allow multiple plugin config blocks to be updated
separately, this adds support for specifying a plugin with an optional
":description" to the name of the plugin block.  For instance, instead of
a single "nginx" block, you can now specify "nginx:local" and
"nginx:remote" to manage two separate blocks that use the same plugin
type.

Since there are now multiple plugin types, some functions were renamed to
better reflect this.  There is also a separate queue for the config
plugins versus the metric plugins.

Since the config might change, we also attempt to clean up old saved
values from previous plugins that are no longer used.

Along these lines, we also now ensure that each instance of a plugin has a
unique name so that results can be stored properly.  This was not
guaranteed before.

This also adds a `skip_newrelic_upload` setting for debugging. If defined
in the config this will simply skip uploading the data to newrelic.