creation of db connection arguments was not thread-safe
discovered we had monitors that were connecting to the wrong database
and causing all kinds of metric issues.  turns out we were using a
global to set the connection argument defaults without copying the
values before modifying them in each thread.

this change also fixes the innodb_data_pending_fsyncs metric so it's
treated as a gauge instead of a counter.